+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Next release
 
+## [0.3.0] - 2021-08-23
+
+* Add `retry` param to error model.
+* Parameters `localizedMessage, description, data, retry` have become named.
+
 ## [0.2.0] - 2021-04-27
 
 * Migrate to null-safety.